Overhaul of Case Reference segment of script.
This now exists within the main script. Just need to sort code to copy to clipboard and tidy up the mess I've made, but the functionality is indeed there. When you load the main script on any case it gives you a nice button which pops a modal, allowing you to grab various tidbits about the case.